+++ src/components/Form/components/FormTableField/FormTableFieldScrollbars/FormTableFieldScrollbars.types.ts
@@ -1,11 +1,7 @@
-import { FormTableFieldContentRefs } from "../FormTableFieldContent";
-
+import type { FormTableFieldContentRefs } from "../FormTableFieldContent";
 import type { RefObject } from "react";
 
 export type FormTableFieldScrollbarsProps = {
-  // scrollerRef: RefObject<HTMLElement>;
-  // contentRef: RefObject<HTMLElement>;
-  // headerRef: RefObject<HTMLElement>;
   contentRefs: RefObject<FormTableFieldContentRefs>;
   scrollerId: string;
 };